rl update

+++ Code/RLClient/Memory.py
@@ -24,7 +24,7 @@ class Memory(object):
 
     def backfill_batch(self, score_mod):
         for idx in range(len(self.memory)):
-            self.memory[idx][-1] += int(score_mod * (((1 - self.DIST_QUOTIENT) / len(self.memory)) * idx + self.DIST_QUOTIENT))
+            self.memory[idx][-1] += float(score_mod * (((1 - self.DIST_QUOTIENT) / len(self.memory)) * idx + self.DIST_QUOTIENT))
 
     def __len__(self):
         return len(self.memory)
